feat!: add faraday 2 support
